Add every transform when recording finger poses.
The previously violated assumption here is that the index will always
match, everywhere. however, a missing finger bone would create an
offset, resulting in hand mangling.

Fix this by always adding a transform, even if it's null. Now the index
continues to line up correctly, even in the absence of some finger
bones.